factor out more tests
