+++ crates/argmin-math/src/faer_tests.rs
@@ -1,3 +1,5 @@
 mod add;
 mod div;
+mod dot;
+mod eye;
 pub mod test_helper;
